backend/requirements.txt
Flask==3.0.0
gunicorn==23.0.0
python-dotenv==1.0.0
ollama

# Azure services (production)
azure-storage-blob==12.19.0
openai>=1.12.0